requirements.txt
asyncio>=3.4.3
bcrypt>=4.0.1
flask>=3.0.0
huggingface_hub>=0.20.2
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.2.0
pinecone-client>=2.2.4
python-dotenv>=1.0.1
pyjwt>=2.8.0
requests>=2.31.0
scikit-learn>=1.4.0
streamlit>=1.30.0
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
firebase-admin
